+++ tests/__init__.py
@@ -28,6 +28,8 @@ CRITICAL = LEVELS[4]
 PYAUD_MODULES = "pyaud.main.plugins"
 SP_OPEN_PROC = "pyaud.utils.Subprocess._open_process"
 OS_GETCWD = "os.getcwd"
+README = Path("README.rst")
+PLUGINS_MODULES_PLUGINS = "plugins.modules.plugins"
 
 
 class NoColorCapsys: